Image_Scripts/img_to_ascii_converter/requirements.txt
Pillow==8.0.0
